+++ src/data_importer/amc_downloaders/helios_holdings/import_all_helios.py
@@ -1,13 +1,13 @@
 """
-src/data_importer/invesco_holdings/import_all_invesco.py
-─────────────────────────────────────────────────────────
-Standalone script to fetch and import Invesco Mutual Fund portfolio holdings into ClickHouse.
+src/data_importer/amc_downloaders/helios_holdings/import_all_helios.py
+───────────────────────────────────────────────────────
+Standalone script to fetch and import Helios Mutual Fund portfolio holdings into ClickHouse.
 
 Usage:
-    python src/data_importer/invesco_holdings/import_all_invesco.py
-    python src/data_importer/invesco_holdings/import_all_invesco.py --full
-    python src/data_importer/invesco_holdings/import_all_invesco.py --dry-run
-    python src/data_importer/invesco_holdings/import_all_invesco.py --month 2026-07
+    python src/data_importer/amc_downloaders/helios_holdings/import_all_helios.py
+    python src/data_importer/amc_downloaders/helios_holdings/import_all_helios.py --full
+    python src/data_importer/amc_downloaders/helios_holdings/import_all_helios.py --dry-run
+    python src/data_importer/amc_downloaders/helios_holdings/import_all_helios.py --month 2026-07
 """
 
 from __future__ import annotations
@@ -23,14 +23,14 @@ sys.path.insert(0, os.getcwd())
 
 from rich.console import Console
 
-from src.data_importer.amc_holdings.importers.invesco import InvescoImporter
+from src.data_importer.amc_holdings.importers.helios import HeliosImporter
 
 console = Console()
 logger = logging.getLogger(__name__)
 
 
 def main() -> None:
-    parser = argparse.ArgumentParser(description="Import Invesco Mutual Fund monthly portfolio holdings")
+    parser = argparse.ArgumentParser(description="Import Helios Mutual Fund monthly portfolio holdings")
     parser.add_argument("--full", action="store_true", help="Ignore watermarks and re-import all historical files")
     parser.add_argument("--dry-run", action="store_true", help="Download and parse files without writing to ClickHouse")
     parser.add_argument("--month", type=str, default="", help="Specific month to import (YYYY-MM, e.g. 2026-07)")
@@ -47,11 +47,11 @@ def main() -> None:
             sys.exit(1)
 
     console.print(
-        f"[bold blue]Starting Invesco Portfolio Import[/bold blue] "
+        f"[bold blue]Starting Helios Portfolio Import[/bold blue] "
         f"(full={args.full}, dry_run={args.dry_run}, target_month={target_month}, fresh={args.fresh})"
     )
 
-    importer = InvescoImporter(
+    importer = HeliosImporter(
         full_reimport=args.full,
         from_year=args.year,
         target_month=target_month,
