+++ src/data_importer/amc_downloaders/invesco_holdings/import_all_invesco.py
@@ -1,13 +1,13 @@
 """
-src/data_importer/canara_holdings/import_all_canara.py
-─────────────────────────────────────────────────────
-Standalone script to fetch and import Canara Robeco Mutual Fund portfolio holdings into ClickHouse.
+src/data_importer/amc_downloaders/invesco_holdings/import_all_invesco.py
+─────────────────────────────────────────────────────────
+Standalone script to fetch and import Invesco Mutual Fund portfolio holdings into ClickHouse.
 
 Usage:
-    python src/data_importer/canara_holdings/import_all_canara.py
-    python src/data_importer/canara_holdings/import_all_canara.py --full
-    python src/data_importer/canara_holdings/import_all_canara.py --dry-run
-    python src/data_importer/canara_holdings/import_all_canara.py --month 2026-07
+    python src/data_importer/amc_downloaders/invesco_holdings/import_all_invesco.py
+    python src/data_importer/amc_downloaders/invesco_holdings/import_all_invesco.py --full
+    python src/data_importer/amc_downloaders/invesco_holdings/import_all_invesco.py --dry-run
+    python src/data_importer/amc_downloaders/invesco_holdings/import_all_invesco.py --month 2026-07
 """
 
 from __future__ import annotations
@@ -23,14 +23,14 @@ sys.path.insert(0, os.getcwd())
 
 from rich.console import Console
 
-from src.data_importer.amc_holdings.importers.canara_robeco import CanaraRobecoImporter
+from src.data_importer.amc_holdings.importers.invesco import InvescoImporter
 
 console = Console()
 logger = logging.getLogger(__name__)
 
 
 def main() -> None:
-    parser = argparse.ArgumentParser(description="Import Canara Robeco Mutual Fund monthly portfolio holdings")
+    parser = argparse.ArgumentParser(description="Import Invesco Mutual Fund monthly portfolio holdings")
     parser.add_argument("--full", action="store_true", help="Ignore watermarks and re-import all historical files")
     parser.add_argument("--dry-run", action="store_true", help="Download and parse files without writing to ClickHouse")
     parser.add_argument("--month", type=str, default="", help="Specific month to import (YYYY-MM, e.g. 2026-07)")
@@ -47,11 +47,11 @@ def main() -> None:
             sys.exit(1)
 
     console.print(
-        f"[bold blue]Starting Canara Robeco Portfolio Import[/bold blue] "
+        f"[bold blue]Starting Invesco Portfolio Import[/bold blue] "
         f"(full={args.full}, dry_run={args.dry_run}, target_month={target_month}, fresh={args.fresh})"
     )
 
-    importer = CanaraRobecoImporter(
+    importer = InvescoImporter(
         full_reimport=args.full,
         from_year=args.year,
         target_month=target_month,
