+++ src/data_importer/amc_downloaders/mirae_holdings/import_all_mirae.py
@@ -1,36 +1,23 @@
 """
-src/data_importer/helios_holdings/import_all_helios.py
-───────────────────────────────────────────────────────
-Standalone script to fetch and import Helios Mutual Fund portfolio holdings into ClickHouse.
-
-Usage:
-    python src/data_importer/helios_holdings/import_all_helios.py
-    python src/data_importer/helios_holdings/import_all_helios.py --full
-    python src/data_importer/helios_holdings/import_all_helios.py --dry-run
-    python src/data_importer/helios_holdings/import_all_helios.py --month 2026-07
+src/data_importer/amc_downloaders/mirae_holdings/import_all_mirae.py
+────────────────────────────────────────────────────
+Stand-alone runner to import Mirae Asset Mutual Fund portfolio disclosures.
 """
 
 from __future__ import annotations
 
 import argparse
-import logging
-import os
 import sys
-from datetime import datetime
-
-# Ensure project root is on sys.path
-sys.path.insert(0, os.getcwd())
-
+from datetime import date, datetime
 from rich.console import Console
 
-from src.data_importer.amc_holdings.importers.helios import HeliosImporter
+from src.data_importer.amc_holdings.importers.mirae import MiraeImporter
 
 console = Console()
-logger = logging.getLogger(__name__)
 
 
 def main() -> None:
-    parser = argparse.ArgumentParser(description="Import Helios Mutual Fund monthly portfolio holdings")
+    parser = argparse.ArgumentParser(description="Import Mirae Asset Mutual Fund monthly portfolio holdings")
     parser.add_argument("--full", action="store_true", help="Ignore watermarks and re-import all historical files")
     parser.add_argument("--dry-run", action="store_true", help="Download and parse files without writing to ClickHouse")
     parser.add_argument("--month", type=str, default="", help="Specific month to import (YYYY-MM, e.g. 2026-07)")
@@ -38,20 +25,23 @@ def main() -> None:
     parser.add_argument("--year", type=int, default=2023, help="Earliest year to consider (default: 2023)")
     args = parser.parse_args()
 
-    target_month = None
+    target_month: date | None = None
     if args.month:
         try:
             target_month = datetime.strptime(args.month, "%Y-%m").date()
         except ValueError:
-            console.print(f"[red]Invalid month format: {args.month}. Expected YYYY-MM.[/red]")
-            sys.exit(1)
+            try:
+                target_month = datetime.strptime(args.month, "%Y-%m-%d").date()
+            except ValueError:
+                console.print(f"[red]Invalid month format: {args.month}. Expected YYYY-MM.[/red]")
+                sys.exit(1)
 
     console.print(
-        f"[bold blue]Starting Helios Portfolio Import[/bold blue] "
+        f"[bold blue]Starting Mirae Asset Portfolio Import[/bold blue] "
         f"(full={args.full}, dry_run={args.dry_run}, target_month={target_month}, fresh={args.fresh})"
     )
 
-    importer = HeliosImporter(
+    importer = MiraeImporter(
         full_reimport=args.full,
         from_year=args.year,
         target_month=target_month,
