--- src/data_importer/axis_holdings/__init__.py
@@ -1,1 +1,0 @@
-"""src/data_importer/axis_holdings"""
