+++ src/scripts/abakkus/import_all_abakkus.py
@@ -17,7 +17,7 @@ import sys
 # Ensure project root is on sys.path
 sys.path.insert(0, os.getcwd())
 
-from src.data_importer.abakkus_holdings.import_all_abakkus import main
+from src.data_importer.amc_downloaders.abakkus_holdings.import_all_abakkus import main
 
 if __name__ == "__main__":
     main()
