+++ src/scripts/axis/import_all_axis.py
@@ -20,7 +20,7 @@ ROOT = Path(__file__).resolve().parents[3]
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
-from src.data_importer.axis_holdings.import_all_axis import main
+from src.data_importer.amc_downloaders.axis_holdings.import_all_axis import main
 
 if __name__ == "__main__":
     main()
