+++ src/scripts/dsp/import_all_dsp_equity.py
@@ -1,4 +1,4 @@
-"""Compat shim — real module is src.data_importer.dsp_holdings.import_all_dsp_equity."""
+"""Compat shim — real module is src.data_importer.amc_downloaders.dsp_holdings.import_all_dsp_equity."""
 import sys as _sys
-import src.data_importer.dsp_holdings.import_all_dsp_equity as _real
+import src.data_importer.amc_downloaders.dsp_holdings.import_all_dsp_equity as _real
 _sys.modules[__name__] = _real
