+++ src/scripts/dsp/import_dsp_history.py
@@ -1,4 +1,4 @@
-"""Compat shim — real module is src.data_importer.dsp_holdings.import_dsp_history."""
+"""Compat shim — real module is src.data_importer.amc_downloaders.dsp_holdings.import_dsp_history."""
 import sys as _sys
-import src.data_importer.dsp_holdings.import_dsp_history as _real
+import src.data_importer.amc_downloaders.dsp_holdings.import_dsp_history as _real
 _sys.modules[__name__] = _real
