+++ src/scripts/dsp/import_dsp_since_2020.py
@@ -1,4 +1,4 @@
-"""Compat shim — real module is src.data_importer.dsp_holdings.import_dsp_since_2020."""
+"""Compat shim — real module is src.data_importer.amc_downloaders.dsp_holdings.import_dsp_since_2020."""
 import sys as _sys
-import src.data_importer.dsp_holdings.import_dsp_since_2020 as _real
+import src.data_importer.amc_downloaders.dsp_holdings.import_dsp_since_2020 as _real
 _sys.modules[__name__] = _real
