+++ src/scripts/dsp/import_latest_dsp.py
@@ -1,4 +1,4 @@
-"""Compat shim — real module is src.data_importer.dsp_holdings.import_latest_dsp."""
+"""Compat shim — real module is src.data_importer.amc_downloaders.dsp_holdings.import_latest_dsp."""
 import sys as _sys
-import src.data_importer.dsp_holdings.import_latest_dsp as _real
+import src.data_importer.amc_downloaders.dsp_holdings.import_latest_dsp as _real
 _sys.modules[__name__] = _real
