+++ src/scripts/helios/import_all_helios.py
@@ -17,7 +17,7 @@ import sys
 # Ensure project root is on sys.path
 sys.path.insert(0, os.getcwd())
 
-from src.data_importer.helios_holdings.import_all_helios import main
+from src.data_importer.amc_downloaders.helios_holdings.import_all_helios import main
 
 if __name__ == "__main__":
     main()
