+++ src/scripts/invesco/import_all_invesco.py
@@ -17,7 +17,7 @@ import sys
 # Ensure project root is on sys.path
 sys.path.insert(0, os.getcwd())
 
-from src.data_importer.invesco_holdings.import_all_invesco import main
+from src.data_importer.amc_downloaders.invesco_holdings.import_all_invesco import main
 
 if __name__ == "__main__":
     main()
